Test loading from installed egg.

+++ openmdao.main/src/openmdao/main/test/test_egg_save.py
@@ -6,6 +6,7 @@ import cPickle
 import logging
 import os
 import shutil
+import subprocess
 import sys
 import unittest
 
@@ -187,7 +188,6 @@ class Oddball(Assembly):
         except RuntimeError, exc:
             if not str(exc).find('empty'):
                 raise exc
-
         try:
             self.debug('method_socket returned %s', self.method_socket())
         except RuntimeError, exc:
@@ -241,7 +241,7 @@ class EggTestCase(unittest.TestCase):
         """ Called after each test in this class. """
         self.model.pre_delete()
         self.model = None
-        if self.egg_name:
+        if self.egg_name and os.path.exists(self.egg_name):
             os.remove(self.egg_name)
         if os.path.exists('Egg'):
             shutil.rmtree('Egg')
@@ -516,10 +516,140 @@ class EggTestCase(unittest.TestCase):
             retcode = self.model.check_save_load()
             self.assertEqual(retcode, 0)
 
+    def test_install_load(self):
+        logging.debug('')
+        logging.debug('test_install_load')
+
+        # Find what is hopefully the correct 'python' command.
+        orig_dir = os.getcwd()
+        python = 'python'
+        if orig_dir.endswith('buildout'):
+            python = os.path.join(orig_dir, 'bin', python)
+        else:
+            index = orig_dir.find('openmdao.main')
+            if index > 0:
+                python = os.path.join(orig_dir[:index],
+                                      'buildout', 'bin', python)
+        logging.debug('Using python: %s' % python)
+
+        # Write to egg.
+        self.egg_name = self.model.save_to_egg()
+
+        install_dir = os.path.join(os.getcwd(), 'install_dir')
+        if os.path.exists(install_dir):
+            shutil.rmtree(install_dir)
+        os.mkdir(install_dir)
+        try:
+            # Create special installer script.
+            installer = os.path.join(install_dir, 'installer.py')
+            out = open(installer, 'w')
+            out.write("""\
+# EASY-INSTALL-ENTRY-SCRIPT: 'setuptools>=0.6c8','console_scripts','easy_install'
+__requires__ = 'setuptools>=0.6c8'
+import sys
+from pkg_resources import load_entry_point
+
+sys.exit(
+   load_entry_point('setuptools>=0.6c8', 'console_scripts', 'easy_install')()
+)
+""")
+            out.close()
+            os.chmod(installer, 0755)
+
+            # Install via subprocess with PYTHONPATH set (for easy_install).
+            logging.debug('Installing via subprocess...')
+            env = os.environ
+            path = env.get('PYTHONPATH', '')
+            if path:
+                path += os.pathsep
+            path += install_dir
+            env['PYTHONPATH'] = path
+            cmdline = '%s %s -d %s %s' % \
+                      (python, installer, install_dir, self.egg_name)
+            stdout = open(os.path.join(install_dir, 'installer.out'), 'w')
+            retcode = subprocess.call(cmdline, env=env, shell=True,
+                                      stdout=stdout, stderr=subprocess.STDOUT)
+            stdout.close()
+            stdout = open(os.path.join(install_dir, 'installer.out'), 'r')
+            for line in stdout:
+                logging.debug('    %s', line.rstrip())
+            stdout.close()
+            self.assertEqual(retcode, 0)
+
+            test_dir = 'EggTest'
+            if os.path.exists(test_dir):
+                shutil.rmtree(test_dir)
+            os.mkdir(test_dir)
+            os.chdir(test_dir)
+            try:
+                # Create load-n-run script.
+                out = open('load-n-run.py', 'w')
+                out.write("""\
+import logging
+import sys
+sys.path.append('%(egg)s')
+import pkg_resources
+
+logging.getLogger().setLevel(logging.DEBUG)
+
+name = '%(name)s'
+
+try:
+    dist = pkg_resources.get_distribution(name)
+except pkg_resources.DistributionNotFound, exc:
+    print 'No distribution found for', exc
+    for path in sys.path:
+        print '   ', path
+    sys.exit(1)
+
+print 'Distribution found.'
+entry_map = dist.get_entry_map()
+for group in entry_map.keys():
+    print '    %%s entry points:' %% group
+    for entry_pt in entry_map[group].values():
+        print '       ', entry_pt
+
+try:
+    loader = dist.load_entry_point('openmdao.components', name)
+except ImportError, exc:
+    print exc
+    sys.exit(2)
+
+print 'Loader found.'
+try:
+    comp = loader()
+except Exception, exc:
+    print exc
+    sys.exit(3)
+
+print 'Running component...'
+comp.run()
+""" % {'egg':os.path.join(install_dir, self.egg_name), 'name':self.model.name})
+                out.close()
+
+                # Load & run in subprocess.
+                logging.debug('Load and run in subprocess...')
+                cmdline = '%s load-n-run.py' % python
+                stdout = open('load-n-run.out', 'w')
+                retcode = subprocess.call(cmdline, env=env, shell=True,
+                                          stdout=stdout,
+                                          stderr=subprocess.STDOUT)
+                stdout.close()
+                stdout = open('load-n-run.out', 'r')
+                for line in stdout:
+                    logging.debug('    %s', line.rstrip())
+                stdout.close()
+                self.assertEqual(retcode, 0)
+
+            finally:
+                os.chdir(orig_dir)
+
+        finally:
+            shutil.rmtree(install_dir)
+
 
 if __name__ == '__main__':
     import nose
-    import sys
     sys.argv.append('--cover-package=openmdao')
     sys.argv.append('--cover-erase')
     nose.runmodule()
